requirements.txt
pyTelegramBotAPI==3.7.3
python-dotenv==0.15.0
Flask==1.1.2